fix(test): fix typescript version conflict (#4)
* chore: move tests inside server (#4)

Keeping tests in a separate package increases the risk of dependency version conflicts, as the test and application environments may diverge.

* perf(ast): cache children nodes (#4)

After resolving the typescript version conflict, performance tests run time increased significantly.
Profiling revealed that `findChildren` accounts for 25% of total test execution time.
This change introduces a simple cache for `findChildren` results to reduce redundant traversals and improve performance. This allows the performance tests to complete in a reasonable amount of time

* fix(ci): fix run commands ordering (#4)